+++ examples/hello_world.py
@@ -1,7 +1,7 @@
 from tau.event import Lambda
-from tau.signal import OneShot
+from tau.signal import From
 from tau.testing import TestSchedulerContextManager
 
 with TestSchedulerContextManager() as scheduler:
-    signal = OneShot(scheduler, ["world"])
+    signal = From(scheduler, ["world"])
     Lambda(scheduler.get_network(), signal, lambda x: print(f"Hello, {x[0].get_value()}!"))