+++ HISTORY.rst
@@ -3,7 +3,14 @@
 Release History
 ---------------
 
+0.0.1 (2020-03-29)
++++++++++++++++++++
+
+- Started harmonizing operators with RX specification
+- Renamed OneShot to From and ForEach to Do
+- Improved documentation
+
 0.0.1 (2020-03-28)
 +++++++++++++++++++
 
-- Hello world.
+- Initial implementation